Display product properties properly on the frontend.
Some of this logic is better placed in a controller, however, by doing
it here we don't have to override the products controller. Ideally, a
draper decorator would fit well here for evaluating
properties_by_category but spree doesn't have it listed as a dependency
and we want to avoid introducing new ones.